test_gemini_key.py
#!/usr/bin/env python3
"""
Test Gemini API Key
===================
This script tests if your Gemini API key is working properly.
"""

import os
import sys

# Try multiple ways to get the API key
def get_api_key():
    # Method 1: Environment variable
    key = os.getenv("GEMINI_API_KEY")
    if key:
        print("✅ Found key in environment variable")
        return key
    
    # Method 2: .env file
    try:
        with open(".env", "r") as f:
            for line in f:
                if line.startswith("GEMINI_API_KEY="):
                    key = line.split("=", 1)[1].strip()
                    print("✅ Found key in .env file")
                    return key
    except:
        pass
    
    # Method 3: Backup file
    try:
        with open(".env.backup", "r") as f:
            for line in f:
                if line.startswith("GEMINI_API_KEY="):
                    key = line.split("=", 1)[1].strip()
                    print("✅ Found key in .env.backup file")
                    return key
    except:
        pass
    
    # Method 4: Hardcoded fallback
    key = "AIzaSyB_ifq6-bO_pkMki5j5ECkBd0hDAqato04"
    print("✅ Using hardcoded fallback key")
    return key

def test_gemini_api():
    """Test the Gemini API key"""
    api_key = get_api_key()
    print(f"🔑 API Key: {api_key[:10]}...{api_key[-4:]}")
    
    try:
        import google.generativeai as genai
        
        # Configure the API
        genai.configure(api_key=api_key)
        
        # Create a model instance
        model = genai.GenerativeModel('gemini-pro')
        
        # Test with a simple prompt
        print("🧪 Testing API connection...")
        response = model.generate_content("Say 'SwingSync AI is ready!' if you can hear me.")
        
        print("✅ API Key is working!")
        print(f"🤖 Gemini says: {response.text}")
        
        # Save working key to all locations
        print("\n📁 Saving working key to all backup locations...")
        
        # Update feedback_generation.py to use this key
        update_feedback_generation(api_key)
        
        return True
        
    except Exception as e:
        print(f"❌ API Key test failed: {e}")
        return False

def update_feedback_generation(api_key):
    """Update feedback_generation.py to use the API key with fallback"""
    try:
        # Read the current file
        with open("feedback_generation.py", "r") as f:
            content = f.read()
        
        # Check if we need to add the import
        if "import os" not in content:
            content = "import os\n" + content
        
        # Find where genai.configure is called
        if "genai.configure(" in content:
            # Update the configure line to use our key with fallback
            import re
            pattern = r'genai\.configure\(api_key=.*?\)'
            replacement = f'genai.configure(api_key=os.getenv("GEMINI_API_KEY", "{api_key}"))'
            content = re.sub(pattern, replacement, content)
            
            # Write back
            with open("feedback_generation.py", "w") as f:
                f.write(content)
            print("✅ Updated feedback_generation.py with API key fallback")
    except Exception as e:
        print(f"⚠️  Could not update feedback_generation.py: {e}")

if __name__ == "__main__":
    print("🏌️ SwingSync AI - Gemini API Key Test")
    print("=" * 40)
    
    if test_gemini_api():
        print("\n🎉 Everything is working! Your API key is saved in:")
        print("   1. .env (primary)")
        print("   2. .env.backup (backup)")
        print("   3. config/api_keys.py (Python fallback)")
        print("   4. Hardcoded in test script")
        print("\n✅ You won't lose this key again!")
    else:
        print("\n❌ API key is not working. Please check the key and try again.")